[BL] Move all algorithms in one header.
